reorganized folderStructure to allow for pip install
