+++ README.md
@@ -1,7 +1,9 @@
 # SpaceNet Utilities
 
-This repository has two python packages, geoTools and evalTools. The geoTools packages is intended to assist in the preprocessing of [SpaceNet](https://spacenetchallenge.github.io/) satellite imagery data corpus hosted on [SpaceNet on AWS](https://aws.amazon.com/public-datasets/spacenet/) to a format that is consumable by machine learning algorithms. The evalTools package is used to evaluate the effectiveness of object detection algorithms using ground truth.
-This is version 2.0 and has been updated with more capabilities to 
+This repository has three python packages, geoTools and evalTools and labelTools. The geoTools packages is intended to assist in the preprocessing of [SpaceNet](https://spacenetchallenge.github.io/) satellite imagery data corpus hosted on [SpaceNet on AWS](https://aws.amazon.com/public-datasets/spacenet/) to a format that is consumable by machine learning algorithms. 
+The evalTools package is used to evaluate the effectiveness of object detection algorithms using ground truth.
+The labelTools package assists in transfering geoJson labels into common label schemes for machine learning frameworks
+This is version 3.0 and has been updated with more capabilities to allow for computer vision applications using remote sensing data
 ## Download Instructions
 
 Further download instructions for the [SpaceNet Dataset](https://github.com/SpaceNetChallenge/utilities/tree/master/content/download_instructions) can be found [here](https://github.com/SpaceNetChallenge/utilities/tree/master/content/download_instructions)
