+++ requirements.txt
@@ -1,3 +1,4 @@
+cython
 numpy
 pandas
 rtree
@@ -7,7 +8,6 @@ geopandas
 rasterio
 scipy
 PIL
-xml
 tqdm
 affine
 pyproj
