+++ spacenetutilities.egg-info/PKG-INFO
@@ -1,0 +1,14 @@
+Metadata-Version: 1.0
+Name: spacenetutilities
+Version: 3.0
+Summary: Geographic pandas extensions
+Home-page: https://github.com/SpaceNetChallenge/utilities
+Author: SpaceNet Contributors
+Author-email: dlindenbaum@iqt.org
+License: APACHE 2
+Description: This repository has three python packages, geoTools and evalTools and labelTools. The geoTools packages is intended to assist in the preprocessing of [SpaceNet](https://spacenetchallenge.github.io/) satellite imagery data corpus hosted on [SpaceNet on AWS](https://aws.amazon.com/public-datasets/spacenet/) to a format that is consumable by machine learning algorithms. 
+        The evalTools package is used to evaluate the effectiveness of object detection algorithms using ground truth.
+        The labelTools package assists in transfering geoJson labels into common label schemes for machine learning frameworks
+        This is version 3.0 and has been updated with more capabilities to allow for computer vision applications using remote sensing data
+        
+Platform: UNKNOWN
