+++ spacenetutilities/scripts/evaluateScenePerformance.py
