+++ spacenetutilities/scripts/splitAOI_Train_Test_Val.py
@@ -1,4 +1,4 @@
-from spaceNetUtilities import labelTools as lT
+from spacenetutilities import labeltools as lT
 import glob
 import argparse
 import random
